+++ packages/sit-onyx/src/components/OnyxListbox/OnyxListbox.vue
@@ -111,12 +111,12 @@ const {
       return i.label.toLowerCase().trim().startsWith(label.toLowerCase());
     });
     if (!firstMatch) return;
-    activeOption.value = firstMatch.id;
+    activeOption.value = firstMatch.value;
   },
 });
 
 const groupedOptions = computed(() => {
-  return props.options.reduce<Record<string, ListboxOption[]>>((acc, currOpt) => {
+  return props.options.reduce<Record<string, ListboxOption<TValue>[]>>((acc, currOpt) => {
     const groupName = currOpt.group ?? "";
     acc[groupName] = acc[groupName] || [];
     acc[groupName].push(currOpt);
@@ -133,7 +133,7 @@ const { vScrollEnd, isScrollEnd } = useScrollEnd({
 const isEmpty = computed(() => props.options.length === 0);
 
 const enabledOptionValues = computed(() =>
-  props.options.filter((i) => !i.disabled).map(({ id }) => id),
+  props.options.filter((i) => !i.disabled).map(({ value }) => value),
 );
 
 /**
@@ -205,31 +205,32 @@ watchEffect(() => {
         </template>
 
         <OnyxListboxOption
-          v-for="option in options as any"
-          :key="option.id.toString()"
+          v-for="option in options"
+          :key="option.value.toString()"
           v-bind="
             headlessOption({
-              value: option.id,
+              value: option.value,
               label: option.label,
               disabled: option.disabled,
               selected:
-                option.id === props.modelValue ||
-                (Array.isArray(props.modelValue) && props.modelValue.includes(option.id)),
+                option.value === props.modelValue ||
+                (Array.isArray(props.modelValue) && props.modelValue.includes(option.value)),
             })
           "
           :multiple="props.multiple"
-          :active="option.id === activeOption"
+          :active="option.value === activeOption"
         >
           {{ option.label }}
         </OnyxListboxOption>
       </ul>
-      <li v-if="props.lazyLoading?.loading" class="onyx-listbox__slot">
-        <OnyxLoadingIndicator class="onyx-listbox__loading" />
-      </li>
 
-      <li v-if="slots.optionsEnd" class="onyx-listbox__slot">
+      <div v-if="props.lazyLoading?.loading" class="onyx-listbox__slot">
+        <OnyxLoadingIndicator class="onyx-listbox__loading" />
+      </div>
+
+      <div v-if="slots.optionsEnd" class="onyx-listbox__slot">
         <slot name="optionsEnd"></slot>
-      </li>
+      </div>
     </div>
     <span v-if="props.message" class="onyx-listbox__message onyx-text--small">
       {{ props.message }}