+++ apps/alpha-test-app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # alpha-test-app
 
+## 0.1.0-alpha.86
+
+### Patch Changes
+
+- Updated dependencies [7951251]
+  - sit-onyx@1.0.0-alpha.89
+
 ## 0.1.0-alpha.85
 
 ### Patch Changes